fix(nonfinite): pin both gather operands to float32 and correct two false claims
The two branches of the distillation-loss metric in SDFT and SDPO gathered different dtypes. One
gathered `torch.zeros(())`, which is always float32; the other gathered `mean_distill_loss`, which
follows the autocast dtype and so is bfloat16 or float16 under mixed precision. `accelerator.gather`
requires every rank to contribute the same dtype and NCCL does not promote, so a run where one rank
took each branch could not complete the collective. Measured under a bfloat16 autocast: the peer's
value arrives as bfloat16 while the zero stays float32. Both sides are now pinned to float32. The
mismatch was introduced by 13a06664 on this branch and is not present on main.

The docstring in tests/distributed/nonfinite_loss.py claimed the assertion "would be false" with a
single rank. It would not. With one rank the sole rank is also the poisoned one, so the assertion
reads 1.0 against an expected 1.0 and passes whether or not the gather is there, which is exactly
the blindness the file exists to remove. The docstring now says that.

Five doc pages head the metric list with "While training and evaluating", but CPO and ORPO record
`frac_nonfinite_loss` with `train_eval="train"`, and NashMD, OnlineDPO and XPO append it inside
`training_step`. Each of those five bullets now states that the trainer records the metric on the
training path only, so readers do not look for it among the evaluation metrics.

+++ trl/experimental/sdft/sdft_trainer.py
@@ -905,10 +905,14 @@ class SDFTTrainer(_BaseTrainer):
             # distillation loss, and a rank that returned here without gathering would leave its next collective
             # paired against a peer's gather of a different quantity. Both are one-element float tensors, so they
             # would complete rather than fail, and each rank would record the other's number. Gather a matching zero
-            # so every rank issues the same collectives whichever branch it takes.
+            # so every rank issues the same collectives whichever branch it takes. Both sides are pinned to
+            # float32, because the peer's value follows the autocast dtype while a bare `torch.zeros` does not,
+            # and a gather whose operands disagree on dtype is not something the collective can reconcile.
             self._log_self_distillation_metric(
                 mode,
-                self.accelerator.gather(torch.zeros((), device=distillation_logits.student_logits.device))
+                self.accelerator.gather(
+                    torch.zeros((), device=distillation_logits.student_logits.device, dtype=torch.float32)
+                )
                 .mean()
                 .item(),
             )
@@ -968,7 +972,7 @@ class SDFTTrainer(_BaseTrainer):
         ).sum() / distillation_logits.loss_mask.sum().clamp(min=1.0)
         self._log_self_distillation_metric(
             mode,
-            self.accelerator.gather(mean_distill_loss).mean().item(),
+            self.accelerator.gather(mean_distill_loss.float()).mean().item(),
         )
         return loss
